Moved templatefuncs to main
